Updates to return the native wasm struct type
