+++ website/src/NewAccount.js
@@ -1,6 +1,6 @@
 import React, {useState} from "react";
 import {Button, Card, Col, Divider, Form, Input, Row} from "antd";
-import {CopyButton} from "./CopyButton";
+import {CopyButton} from "./components/CopyButton";
 import {useAleoWASM} from "./aleo-wasm-hook";
 
 export const NewAccount = () => {
